rebased and merge conflicts resolved
