+++ assets/js/animation.js
@@ -1,0 +1,7 @@
+var animation = bodymovin.loadAnimation({
+  container: document.getElementById('animated-icon'),
+  renderer: 'svg',
+  loop: true,
+  autoplay: true,
+  path: "./assets/js/icon.json"
+})
