+++ index.html
@@ -17,9 +17,9 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/bodymovin/5.7.4/lottie.min.js"
         integrity="sha512-m0RQU4SBx0p/bLwRiI4fJBGRafVLZ4s86wRe1+OAx5EXbcWiS/X1jvYdJQRD8jOoIVl+WTyVeMawUWVCh1O8+Q=="
         crossorigin="anonymous"></script>
-    <script src="../static/assets/js/animation.js"></script>
+    <script src="./assets/js/animation.js"></script>
     <!--Custom CSS-->
-    <link rel="stylesheet" href="../static/assets/css/styles.css">
+    <link rel="stylesheet" href="./assets/css/styles.css">
     <title>DSC Afterhours</title>
 </head>
 
@@ -42,7 +42,7 @@
             <div class="mob-bars">
                 <span>
                     <a href="#home">
-                        <img src="../static/assets/img/DSCWhitedsc.svg" alt="dsc" class="nav-logo">
+                        <img src="./assets/img/DSCWhitedsc.svg" alt="dsc" class="nav-logo">
                     </a>
                 </span>
                 <img src="./assets/icons/Vectorbars.svg" alt="menu" class="mobile-bars">
@@ -73,15 +73,15 @@
                 </p>
                 <div class="home-icons">
                     <a href="https://open.spotify.com/show/3N03C8MuIQPSqqGZp8XF56">
-                        <img src="../static/assets/icons/Vectoricon2.svg" alt="social" class="home-icon"></a>
+                        <img src="./assets/icons/Vectoricon2.svg" alt="social" class="home-icon"></a>
                     <a href="https://podcasts.apple.com/us/podcast/id1536768786">
-                        <img src="../static/assets/icons/Vectoricon1.svg" alt="social" class="home-icon"></a>
+                        <img src="./assets/icons/Vectoricon1.svg" alt="social" class="home-icon"></a>
                     <a href="https://pca.st/m1y1rnmy">
-                        <img src="../static/assets/icons/Vectoricon3.svg" alt="social" class="home-icon"></a>
+                        <img src="./assets/icons/Vectoricon3.svg" alt="social" class="home-icon"></a>
                     <a href="https://podcasts.google.com/feed/aHR0cHM6Ly9hbmNob3IuZm0vcy8zYzU1YTVmMC9wb2RjYXN0L3Jzcw==">
-                        <img src="../static/assets/icons/Vectoricon4.svg" alt="social" class="home-icon"></a>
+                        <img src="./assets/icons/Vectoricon4.svg" alt="social" class="home-icon"></a>
                     <a href="https://overcast.fm/itunes1536768786/dsc-afterhours">
-                        <img src="../static/assets/icons/Vectoricon6.svg" alt="social" class="home-icon"></a>
+                        <img src="./assets/icons/Vectoricon6.svg" alt="social" class="home-icon"></a>
                     <a href="https://radiopublic.com/dsc-afterhours-WoL3jR">
                         <img src="./assets/icons/Vectoricon5.svg" alt="social" class="home-icon"></a> </div>
 
@@ -91,7 +91,7 @@
     </section>
     <section id="aboutus" class="section">
         <div class="about">
-            <img src="../static/assets/img/group_dsc.png" alt="group" class="group">
+            <img src="./assets/img/group_dsc.png" alt="group" class="group">
             <div class="about-text">
                 <h1>
                     About Us
@@ -124,8 +124,8 @@
 
     <section id="footer">
         <div class="footer-one">
-            <img src="../static/assets/img/full-logo.svg" alt="dsc" class="full-logo">
-            <img src="../static/assets/img/full-logo-center.svg" alt="dsc" class="full-logo-center">
+            <img src="./assets/img/full-logo.svg" alt="dsc" class="full-logo">
+            <img src="./assets/img/full-logo-center.svg" alt="dsc" class="full-logo-center">
             <p class="mwl">Made with <i class="far fa-heart"></i> by DSC VIT</p>
 
         </div>
