--- static/assets/js/animation.js
@@ -1,7 +1,0 @@
-var animation = bodymovin.loadAnimation({
-  container: document.getElementById('animated-icon'),
-  renderer: 'svg',
-  loop: true,
-  autoplay: true,
-  path: "../static/assets/js/icon.json"
-})
